% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_ardl.R
\name{.aardl_bootstrap}
\alias{.aardl_bootstrap}
\title{AARDL Bootstrap Procedure}
\usage{
.aardl_bootstrap(dy, design, n_level, k, nboot, idx_h0 = NULL)
}
\description{
AARDL Bootstrap Procedure
}
\keyword{internal}
