% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_ardl.R
\name{.aardl_stats}
\alias{.aardl_stats}
\title{Statistics of aardl(): F_pss (Fov), t_dep (t) and F_ind (Find)}
\usage{
.aardl_stats(dy, design, k, case)
}
\description{
Statistics of aardl(): F_pss (Fov), t_dep (t) and F_ind (Find)
}
\keyword{internal}
