% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.boot_ardl_stats}
\alias{.boot_ardl_stats}
\title{Statistics of boot_ardl on a (bootstrap) sample}
\usage{
.boot_ardl_stats(y, X, x_vars, p, q, case)
}
\description{
Fov, t and Find with the design of boot_ardl().
}
\keyword{internal}
