% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.bootstrap_bounds}
\alias{.bootstrap_bounds}
\title{Bootstrap Bounds Test}
\usage{
.bootstrap_bounds(ardl_data, model_ur, case, k, nboot, parallel, ncores)
}
\description{
Bootstrap Bounds Test
}
\keyword{internal}
