% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.bounds_conclusion}
\alias{.bounds_conclusion}
\title{Bounds Test Conclusion}
\usage{
.bounds_conclusion(F_stat, t_stat, cv_F, cv_t, k, case)
}
\description{
Bounds Test Conclusion
}
\keyword{internal}
