% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_ardl.R
\name{.breusch_godfrey_test}
\alias{.breusch_godfrey_test}
\title{Breusch-Godfrey Serial Correlation Test}
\usage{
.breusch_godfrey_test(resid, order = 2)
}
\description{
Breusch-Godfrey Serial Correlation Test
}
\keyword{internal}
