% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_ardl.R
\name{.breusch_pagan_test}
\alias{.breusch_pagan_test}
\title{Breusch-Pagan Heteroskedasticity Test}
\usage{
.breusch_pagan_test(model)
}
\description{
Breusch-Pagan Heteroskedasticity Test
}
\keyword{internal}
