% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.compute_F_overall}
\alias{.compute_F_overall}
\title{Compute Overall F-statistic}
\usage{
.compute_F_overall(model_ur, model_r, k, case)
}
\description{
Compute Overall F-statistic
}
\keyword{internal}
