% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.compute_F_stat}
\alias{.compute_F_stat}
\title{Compute F-statistic for Bounds Test}
\usage{
.compute_F_stat(model_ur, model_r, k, case)
}
\description{
Compute F-statistic for Bounds Test
}
\keyword{internal}
