% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augmented_ardl.R
\name{.decompose_asymmetric}
\alias{.decompose_asymmetric}
\title{Decompose Variables for Asymmetric Analysis}
\usage{
.decompose_asymmetric(X, threshold = 0)
}
\description{
Decompose Variables for Asymmetric Analysis
}
\keyword{internal}
