% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_ardl.R
\name{.detect_ardl_breaks}
\alias{.detect_ardl_breaks}
\title{Detect Structural Breaks}
\usage{
.detect_ardl_breaks(F_stats, ec_coefs, lr_coefs)
}
\description{
Detect Structural Breaks
}
\keyword{internal}
