% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_ardl.R
\name{.detect_fourier_breaks}
\alias{.detect_fourier_breaks}
\title{Detect Fourier Structural Breaks}
\usage{
.detect_fourier_breaks(fourier_coefs, n, k)
}
\description{
Detect Fourier Structural Breaks
}
\keyword{internal}
