% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.estimate_ardl_restricted}
\alias{.estimate_ardl_restricted}
\title{Estimate Restricted ARDL Model}
\usage{
.estimate_ardl_restricted(ardl_data, case)
}
\description{
Estimate Restricted ARDL Model
}
\keyword{internal}
