% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.estimate_ardl_unrestricted}
\alias{.estimate_ardl_unrestricted}
\title{Estimate Unrestricted ARDL Model}
\usage{
.estimate_ardl_unrestricted(ardl_data, case)
}
\description{
Estimate Unrestricted ARDL Model
}
\keyword{internal}
