% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_ardl.R
\name{.estimate_ardl_window}
\alias{.estimate_ardl_window}
\title{Estimate ARDL for Single Window}
\usage{
.estimate_ardl_window(data, y_var, x_vars, p, q, case)
}
\description{
Estimate ARDL for Single Window
}
\keyword{internal}
