% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ardl.R
\name{.estimate_dfe}
\alias{.estimate_dfe}
\title{Dynamic Fixed Effects Estimator}
\usage{
.estimate_dfe(data, y_var, x_vars, id, time, p, q, trend)
}
\description{
Internal function for DFE estimation
}
\keyword{internal}
