% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_ardl.R
\name{.estimate_fourier_ardl}
\alias{.estimate_fourier_ardl}
\title{Internal Fourier ARDL Estimation}
\usage{
.estimate_fourier_ardl(data, y_var, x_vars, p, q, k, case)
}
\description{
Internal Fourier ARDL Estimation
}
\keyword{internal}
