% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ardl.R
\name{.estimate_mg_internal}
\alias{.estimate_mg_internal}
\title{Internal MG estimation}
\usage{
.estimate_mg_internal(group_data, groups, k_x, trend)
}
\description{
Internal MG estimation
}
\keyword{internal}
