% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ardl.R
\name{.estimate_pmg}
\alias{.estimate_pmg}
\title{Pooled Mean Group Estimator}
\usage{
.estimate_pmg(data, y_var, x_vars, id, time, p, q, trend, maxiter, tol)
}
\description{
Internal function for PMG estimation
}
\keyword{internal}
