% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnardl.R
\name{.estimate_qnardl_tau}
\alias{.estimate_qnardl_tau}
\title{Estimate QNARDL for Single Quantile}
\usage{
.estimate_qnardl_tau(model_data, y_col, x_cols, tau)
}
\description{
Estimate QNARDL for Single Quantile
}
\keyword{internal}
