% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnardl.R
\name{.extract_long_run}
\alias{.extract_long_run}
\title{Extract Long-Run Coefficients}
\usage{
.extract_long_run(results_by_tau, level_vars, tau)
}
\description{
Extract Long-Run Coefficients
}
\keyword{internal}
