% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.panel_nardl_decompose}
\alias{.panel_nardl_decompose}
\title{Panel NARDL Decomposition}
\usage{
.panel_nardl_decompose(data, x_vars, id, time, threshold)
}
\description{
Panel NARDL Decomposition
}
\keyword{internal}
