% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.pnardl_bootstrap}
\alias{.pnardl_bootstrap}
\title{Panel NARDL Bootstrap}
\usage{
.pnardl_bootstrap(
  data,
  y_var,
  x_vars,
  id,
  time,
  p,
  q,
  estimator,
  effect,
  nboot,
  groups
)
}
\description{
Panel NARDL Bootstrap
}
\keyword{internal}
