% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.pnardl_dfe}
\alias{.pnardl_dfe}
\title{Panel NARDL DFE Estimator}
\usage{
.pnardl_dfe(data, y_var, x_vars, orig_vars, id, time, p, q, effect, k)
}
\description{
Panel NARDL DFE Estimator
}
\keyword{internal}
