% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.pnardl_mg}
\alias{.pnardl_mg}
\title{Panel NARDL MG Estimator}
\usage{
.pnardl_mg(data, y_var, x_vars, orig_vars, id, time, p, q, effect, k, groups)
}
\description{
Panel NARDL MG Estimator
}
\keyword{internal}
