% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.pnardl_pmg}
\alias{.pnardl_pmg}
\title{Panel NARDL PMG Estimator}
\usage{
.pnardl_pmg(data, y_var, x_vars, orig_vars, id, time, p, q, effect, k)
}
\description{
Panel NARDL PMG Estimator
}
\keyword{internal}
