% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_ardl.R
\name{.prepare_ardl_data}
\alias{.prepare_ardl_data}
\title{Prepare ARDL Data}
\usage{
.prepare_ardl_data(data, y_var, x_vars, time, p, q, trend)
}
\description{
Internal function to prepare data for ARDL estimation
}
\keyword{internal}
