% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.prepare_ardl_ts_data}
\alias{.prepare_ardl_ts_data}
\title{Prepare Time Series Data for ARDL}
\usage{
.prepare_ardl_ts_data(data, y_var, x_vars, p, q, case)
}
\description{
Prepare Time Series Data for ARDL
}
\keyword{internal}
