% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fourier_ardl.R
\name{.prepare_fourier_data}
\alias{.prepare_fourier_data}
\title{Prepare Fourier ARDL Data}
\usage{
.prepare_fourier_data(data, y_var, x_vars, p, q, k, case)
}
\description{
Prepare Fourier ARDL Data
}
\keyword{internal}
