% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qnardl.R
\name{.prepare_qnardl_data}
\alias{.prepare_qnardl_data}
\title{Prepare QNARDL Data with Partial Sum Decomposition}
\usage{
.prepare_qnardl_data(data, y_var, x_vars, decompose, p, q, trend)
}
\description{
Prepare QNARDL Data with Partial Sum Decomposition
}
\keyword{internal}
