% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{.pss_pvalues}
\alias{.pss_pvalues}
\title{PSS bounds test p-values (Kripfganz and Schneider, 2020)}
\usage{
.pss_pvalues(F_stat, t_stat, k, case, n = NULL, sr = 0)
}
\description{
PSS bounds test p-values (Kripfganz and Schneider, 2020)
}
\keyword{internal}
