% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/panel_nardl.R
\name{.stack_lr_data}
\alias{.stack_lr_data}
\title{Stack Long-Run Data for PMG}
\usage{
.stack_lr_data(data, y_var, x_vars, id, time, phi, sr_results, groups)
}
\description{
Stack Long-Run Data for PMG
}
\keyword{internal}
