% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro_data}
\alias{macro_data}
\title{Time Series Macroeconomic Data}
\format{
A data frame with 100 rows and 5 variables: quarter, gdp, inflation,
  investment, trade.
}
\source{
Simulated data for demonstration purposes
}
\usage{
macro_data
}
\description{
A simulated time series dataset of macroeconomic variables
for a single country over 100 quarters.
}
\examples{
data(macro_data)
head(macro_data)

}
\keyword{datasets}
