% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{plot.boot_ardl}
\alias{plot.boot_ardl}
\title{Plot Bootstrap Distribution}
\usage{
\method{plot}{boot_ardl}(x, which = "both", ...)
}
\arguments{
\item{x}{An object of class "boot_ardl"}

\item{which}{Character. "F", "t", or "both" (default)}

\item{...}{Additional arguments passed to plotting functions}
}
\description{
Plot the bootstrap distribution of test statistics
}
