% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbnardl.R
\name{plot.fbnardl}
\alias{plot.fbnardl}
\title{Plot method for fbnardl objects}
\usage{
\method{plot}{fbnardl}(x, type = c("multipliers", "kstar"), ...)
}
\arguments{
\item{x}{An object of class \code{"fbnardl"}}

\item{type}{\code{"multipliers"} (dynamic multipliers with bands) or
\code{"kstar"} (sum of squared residuals by Fourier frequency).}

\item{...}{Not used}
}
\description{
Plot method for fbnardl objects
}
