% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{print.boot_ardl}
\alias{print.boot_ardl}
\title{Print method for boot_ardl}
\usage{
\method{print}{boot_ardl}(x, ...)
}
\arguments{
\item{x}{An object of class "boot_ardl"}

\item{...}{Not used}
}
\description{
Print method for boot_ardl
}
