% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_ardl.R
\name{summary.boot_ardl}
\alias{summary.boot_ardl}
\title{Summary method for boot_ardl}
\usage{
\method{summary}{boot_ardl}(object, ...)
}
\arguments{
\item{object}{An object of class "boot_ardl"}

\item{...}{Not used}
}
\description{
Summary method for boot_ardl
}
