% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{.mtnardl_bounds_conclusion}
\alias{.mtnardl_bounds_conclusion}
\title{MT-NARDL Bounds Test Conclusion}
\usage{
.mtnardl_bounds_conclusion(F_stat, t_stat, cv, boot = NULL)
}
\description{
MT-NARDL Bounds Test Conclusion
}
\keyword{internal}
