% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{.select_optimal_thresholds}
\alias{.select_optimal_thresholds}
\title{Select Optimal Thresholds}
\usage{
.select_optimal_thresholds(y, X, n_thresholds, p, q, case)
}
\description{
Select Optimal Thresholds
}
\keyword{internal}
