% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtnardl.R
\name{.test_regime_asymmetry}
\alias{.test_regime_asymmetry}
\title{Test Regime Asymmetry}
\usage{
.test_regime_asymmetry(
  coefs,
  vcov_mat,
  x_vars,
  regime_names,
  thresholds,
  k_total
)
}
\description{
Test Regime Asymmetry
}
\keyword{internal}
